ReportSubtotal-package {ReportSubtotal}	R Documentation

ReportSubtotal: Adds Subtotals to Data Reports

Description

Adds subtotal rows / sections (a la SAS's Proc Tabulate) to a Group By output by running a series of Group Bys with lesser sets of variables and combining the results with the original.

Author(s)

Maintainer: Yoni Aboody yoniaboody@gmail.com [copyright holder]

[Package ReportSubtotal version 0.0.0.9900 Index]