subtotal_dupe_removal {ReportSubtotal}	R Documentation

Duplicate Subtotal Row Removal Function

Description

If a data report with subtotals (such as a Group By + subtotal output) has some sub-sections with only one row / one set of rows, then the subtotal function will produce row(s) totalling one row (each). These are duplicate subtotal rows, which this function removes.

Usage


subtotal_dupe_removal(data, column, iterator = 2, skip = 0, remove = "All")

Arguments

data
A data frame or tibble

column
The column containing duplicate row labels

iterator
The minimum number of rows meant to be between each section. Two if removing duplicate rows.

skip
The number of rows to skip removing rows from. Usually zero. Use to dodge NA values.

remove
The subtotal label. Usually "All".

Details

A lead() version of the parameter column is added, which lines up the parameter column with its observations x rows deeper (x = iterator). If both columns equal y value (y = remove) then there are only (x / 2) rows between them, making the row a duplicate subtotal row to be removed. This doesn't work for the last few rows, so they're covered with another logical statement using our index column. Those two columns are removed before the output.

Value

The data report without duplicate subtotal rows.

Examples


library(dplyr)

group_by(mtcars, cyl, gear, vs) %>% summarise(Sum = sum(disp)) %>%
subtotal_row(mtcars, "disp") %>%
subtotal_dupe_removal(3, skip = 3)

[Package ReportSubtotal version 0.0.0.9000 Index]