subtotal_section {ReportSubtotal}	R Documentation

Subtotal Section Addition Function

Description

Add subtotal sections to a data report.

Usage


subtotal_section(
  report,
  frame,
  vars = "Population",
  aggregator = "sum",
  parameters = list(),
  subtotal_label = "All",
  exclude = numeric(0)
)

Arguments

report
The data report.

frame
The data frame used to generate the data report.

vars
The names of the column(s) in the data frame to be aggregated.

aggregator
The aggregation function(s) to be used.

parameters
The parameter(s) to be used with the aggregation function(s).

subtotal_label
The label to be used for subtotal rows.

exclude
The variables which need only subtotal rows and not sections.

Details

The dataset and report are factorized, and a series of reports with the same variables are then generated, but with some variables replaced by a subtotal label, which concentrates all sections related to those variables into one subtotal row per section (or one section per many sections). Those subtotal reports are then combined with the original report, and then sorted (with the subtotal label as the first factor level).

Value

The data report with subtotal sections included.

Examples


library(dplyr)

group_by(iris, Species) %>%
summarise(Petal_Sum = sum(Petal.Width)) %>%
subtotal_section(iris, "Petal.Width")

group_by(mtcars, cyl, gear) %>%
summarise(Average_Miles = mean(mpg)) %>%
subtotal_section(mtcars, "mpg", aggregator = "mean")

[Package ReportSubtotal version 0.0.0.9000 Index]